Curr Sens: clean example, make a lof of refactorings
