+++ examples/current-sensor/src/GUI/GUI.cpp
@@ -2,7 +2,8 @@
 #include <vector>
 #include <functional>
 #include <memory>
-#include "WindowPack.h"
+
+
 using namespace emGUI;
 
 static xStatusBar * statusbar;
@@ -12,87 +13,9 @@ static xButton * crossButton;
 
 bool MainWindowCloseRequestHdl(xWidget *);
 
-class TPTest: public DisposableWindow<WINDOW_TP_TEST, TPTest> {
-public:
-  void create() {
-    
-    vWindowSetFullScreen(xThis, true);
 
-    vWidgetSetClickable(xThis, true);
-    xThis->pxCheckTSRoutine = [] (xWidget *pxW, xTouchEvent_t * xEv) -> bool {
-      auto x = xEv->x;
-      auto y = xEv->y;
-      auto rectSize = 2;
-      x = (x >= rectSize)?x-rectSize:x;
-      y = (y >= rectSize)?y-rectSize:y;
-      pxDrawHDL()->vRectangle(x, y, x + rectSize, y + rectSize, 0, true);
-      return false;
-    };
-  }
-  bool onDrawUpdate() { 
-    pxDrawHDL()->vHLine(0, EMGUI_LCD_HEIGHT/2, EMGUI_LCD_WIDTH, EMGUI_COLOR_GRAY);
-    pxDrawHDL()->vVLine(EMGUI_LCD_WIDTH/2, 0,  EMGUI_LCD_HEIGHT, EMGUI_COLOR_GRAY);   
-    return true; 
-  }
-  uint16_t prvGetNextColor() {
-    static std::vector<uint16_t> colors = {
-      EMGUI_COLOR_WHITE,
-      EMGUI_COLOR_RED,
-      EMGUI_COLOR_GREEN,
-      EMGUI_COLOR_BLUE,
-      EMGUI_COLOR_BLACK
-
-    };
-    
-    if (++colorNum >= colors.size())
-      colorNum = 0;
-    return colors[colorNum];
-  }
-
-    bool onKeypress(uint16_t uEv) {
-    if (uEv == 0) {
-      vWidgetSetBgColor(xThis, prvGetNextColor(), false);
-      return true;
-    }
-
-    return false;
-  }
-
-protected:
-  unsigned int colorNum = 0;
-};
-
-class WindowMain : public DisposableWindow<WINDOW_MENU, WindowMain> {
-public:
-  void create() {
-    vWindowSetHeader(xThis, "Main menu");
-    uint8_t offset = 20;
-    uint8_t btnWidth = 120;
-    uint8_t btnHeight = 120;
-    uint8_t row1 = offset;
-    uint8_t column1 = offset;
-    uint8_t column2 = EMGUI_LCD_WIDTH - offset - btnWidth;
-
-	auto paintBtn = pxButtonCreateFromText(column1, row1, btnWidth, btnHeight, "Paint", xThis);
-	vButtonSetOnClickHandler(paintBtn,
-		[](xWidget *) {
-      TPTest::getInstance()->open();
-      return true;
-    });
-  auto plotBtn = pxButtonCreateFromText(column2, row1, btnWidth, btnHeight, "Plot", xThis);
-  vButtonSetOnClickHandler(plotBtn, 
-    [](xWidget *) {
-      WindowPlot::getInstance()->open();
-      return true;
-    });
-  }
-
-  bool onCloseRequest() {
-    return false; //allow window close
-  }
-};
-
-// Action on interface creatings
+// Action on interface creating
+// Init statusbar and open main window
 bool bGUIonWindowManagerCreateHandler(xWidget *) {
   vWindowManagerSetKeypressHandler([](int iID, uint16_t uEv){
 	  (void)iID;
@@ -110,6 +33,7 @@ bool bGUIonWindowManagerCreateHandler(xWidget *) {
 
   header = std::make_unique<WindowHeader>(usX, usY, usW, 0, getSmallFont(), EMGUI_WINDOW_HEADER_LENGTH, statusbar);
 
+  // You should use R5G6B5 bmp (just put it in SPIFFS) I prefer to use GIMP to convert bmp. 
   usX = usWidgetGetW(statusbar) - pxDrawHDL()->usGetPictureW("/cross.bmp") - 1 ;
   usY = 1;
 
@@ -126,8 +50,4 @@ bool bGUIonWindowManagerCreateHandler(xWidget *) {
 
 void GUIInit(){
 	pxWindowManagerCreate(bGUIonWindowManagerCreateHandler);
-}
-
-void vGUIHandlePeriph(eHW event, int32_t param1) {
-
 }
