+++ examples/current-sensor/src/GUI/WindowPack.h
@@ -1,5 +1,4 @@
-#ifndef _WINDOW_PACK_
-#define _WINDOW_PACK_
+#pragma once
 
 
 //#include "emGUI/emGUI.hpp"
@@ -10,21 +9,15 @@
 
 typedef enum {
 	WINDOW_MENU,
-	WINDOW_TP_TEST,
+	WINDOW_PAINT,
 	WINDOW_PLOT
 }eWindows;
-
-#define COLOR_BLUE EMGUI_COLOR_SEA_3
-#define so 5
-#define bo 14
 
 xFont getBigFont();
 xFont getSmallFont();
 xFont getBoldFont();
 
 // #include "TimeClock.hpp"
-#include "ButtonWithBg.hpp"
-#include "TextFrame.hpp"
 #include "WindowPlot.h"
-
-#endif // !_WINDOW_PACK_
+#include "WindowPaint.h"
+#include "WindowMain.h"
